move observations into wrapper
