+++ nidup/pysc2/commanders.py
@@ -1,5 +1,5 @@
 
-from nidup.pysc2.observations import Observations
+from nidup.pysc2.wrapper.observations import Observations
 from nidup.pysc2.order import Order
 from nidup.pysc2.information import BaseLocation, StepIndex
 from nidup.pysc2.production.build import OrdersSequence, CenterCameraOnCommandCenter, BuildSupplyDepot, BuildFactory, BuildRefinery, BuildBarracks, BuildTechLabBarracks, MorphOrbitalCommand
