+++ nidup/pysc2/information.py
@@ -1,6 +1,6 @@
 
 
-from nidup.pysc2.observations import Observations, ScreenFeatures
+from nidup.pysc2.wrapper.observations import Observations, ScreenFeatures
 from nidup.pysc2.wrapper.unit_types import UnitTypeIds
 
 # Parameters
