+++ nidup/pysc2/smart_agents.py
@@ -2,7 +2,7 @@ import math
 import numpy as np
 from nidup.pysc2.learning.qlearning import QLearningTable, QLearningTableStorage
 from nidup.pysc2.learning.game_results import GameResultsTable
-from nidup.pysc2.observations import Observations
+from nidup.pysc2.wrapper.observations import Observations
 from nidup.pysc2.smart_orders import Location, BuildBarracks, BuildSupplyDepot, BuildMarine, Attack, NoOrder
 from nidup.pysc2.wrapper.unit_types import UnitTypeIds
 from pysc2.agents.base_agent import BaseAgent
